Replacing translations; renamed properties
I was working on the QML but then I got distracted by how they were abusing i18n.
I probably should have committed after I renamed all the properties (and made them properties so they're easier to access in the QML) but if that's the problem it should be fairly easy to diff out.