Refactored .py file, removed .pot workflow
Renamed the hell out of stuff in the .py file so even if I don't understand half the logic I might at least know what to look for. The formatting is still a mess in a lot of places but that's the lesser of the problems.
Removed the update_pot workflow since I have my own scripts for building the language files.
Updated the package build workflow so it only contains the necessary files.